Improved Bet Class
The Bet class now stores bet data in a static variable (__static_data__) and automatically updates it everytime a bet object is created.
I also added 2 static methods, reset() and betWithMajority.

betWithMajority is the remake of majorityBet and still returns the same output.

reset resets the __static_data__ variable to {"BLUE":0, "RED":0}

I also added a class method named *extractBet* which is the remake of betExtract with an implementation of logBet (which is a remake of logData) and f-strings